+++ docs/position-selector/demo/demo-01.tsx
@@ -1,6 +1,14 @@
-import React from 'react';
-import { PositionSelector } from '@sensoro/sensoro-map';
+import React from "react";
+import { PositionSelector } from "@sensoro/sensoro-map";
 
 export default () => {
-  return <PositionSelector small style={{ width: 800, height: 400 }} />;
+  return (
+    <PositionSelector
+      value={{ lnglat: [116.378517, 39.865246] }}
+      onChange={(val) => {
+        console.info("=====>坐标change", val);
+      }}
+      style={{ width: 600, height: 400 }}
+    />
+  );
 };
