+++ docs/position-selector/demo/demo-04.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
-import Icon from '@sensoro/sensoro-design/es/icon';
-import { Position } from '@sensoro/sensoro-map';
+import { PositionSelector } from '@sensoro/sensoro-map';
 
 export default () => {
   return (
-    <Position
+    <PositionSelector
       style={{ width: 550, height: 300 }}
-      value={{ lnglat: [116.905163, 40.006047], location: '' }}
-      icon={<Icon type="icon-car-outlined" style={{ fontSize: 24, color: 'red' }} />}
+      value={{ lnglat: [116.378517, 39.865246], location: '' }}
+      isReadOnly
     />
   );
 };
