+++ docs/position-selector/demo/demo-05.tsx
@@ -1,19 +1,13 @@
 import React from 'react';
 import Icon from '@sensoro/sensoro-design/es/icon';
-import { Position } from '@sensoro/sensoro-map';
+import { PositionSelector } from '@sensoro/sensoro-map';
 
 export default () => {
   return (
-    <Position
+    <PositionSelector
       style={{ width: 550, height: 300 }}
       value={{ lnglat: [116.905163, 40.006047], location: '' }}
-      marker={{
-        render: () => {
-          return (
-            <Icon type="icon-car-outlined" style={{ fontSize: 24, color: 'red' }} />
-          )
-        }
-      }}
+      icon={<Icon type="icon-car-outlined" style={{ fontSize: 24, color: 'red' }} />}
     />
   );
 };
